AG-17329 Override pickNodesInBBoxPredicate in sunburst series
The base implementation was entering the `!(node instanceof Sector)` branch on
some Group scene-nodes, with was causing some buggy hit-testing results.
