+++ packages/ag-charts-community/src/chart/series/polar/pieUtil.ts
@@ -1,10 +1,15 @@
-import type { InternalAgColorType, Point } from 'ag-charts-core';
-import { isBetweenAngles, toRadians } from 'ag-charts-core';
+import type { BoxBounds, InternalAgColorType, Point } from 'ag-charts-core';
+import { boxContains, isBetweenAngles, toRadians } from 'ag-charts-core';
+import type { AgSelectionContainment } from 'ag-charts-types';
 
 import type { FromToMotionPropFn, FromToMotionPropFnContext, NodeUpdateState } from '../../../motion/fromToMotion';
-import type { Sector } from '../../../scene/shape/sector';
+import type { Group } from '../../../scene/group';
+import type { Node } from '../../../scene/node';
+import { Sector } from '../../../scene/shape/sector';
+import { Transformable } from '../../../scene/transformable';
+import { boxCollidesSector } from '../../../scene/util/sector';
 import type { Marker } from '../../marker/marker';
-import type { SeriesNodePickMatch } from '../series';
+import type { PickNodesInBBoxPredicate, SeriesNodePickMatch } from '../series';
 
 type AnimatableSectorDatum = {
     radius: number;
@@ -164,4 +169,35 @@ export function pickByMatchingAngle(series: SectorSeries, point: Point): SeriesN
         }
     }
     return undefined;
+}
+
+export function pickSectorsInBBoxPredicate(series: {
+    properties: { selection: { containment: AgSelectionContainment } };
+    contentGroup: Group;
+}): PickNodesInBBoxPredicate {
+    const unreachable = (a: never): never => a;
+    const containment = series.properties.selection.containment;
+    switch (containment) {
+        case 'any':
+            return (selectionBox: BoxBounds, node: Node<unknown>): boolean => {
+                if (node instanceof Sector) {
+                    const offset = Transformable.fromCanvasPoint(series.contentGroup, selectionBox.x, selectionBox.y);
+                    const seriesSelectionBox: BoxBounds = {
+                        x: offset.x,
+                        y: offset.y,
+                        width: selectionBox.width,
+                        height: selectionBox.height,
+                    };
+                    return boxCollidesSector(seriesSelectionBox, node);
+                }
+                return false;
+            };
+        case 'all':
+            return (selectionBox: BoxBounds, node: Node<unknown>): boolean => {
+                const { x, y, width, height } = Transformable.toCanvas(series.contentGroup, node.getBBox());
+                return boxContains(selectionBox, x, y, width, height);
+            };
+        default:
+            return unreachable(containment);
+    }
 }
