+++ packages/ag-charts-community/src/chart/series/series.ts
@@ -65,9 +65,7 @@ import { Group, TranslatableGroup } from '../../scene/group';
 import { type Node, PointerEvents } from '../../scene/node';
 import type { Selection } from '../../scene/selection';
 import type { Path } from '../../scene/shape/path';
-import { Sector } from '../../scene/shape/sector';
 import { Transformable } from '../../scene/transformable';
-import { boxCollidesSector } from '../../scene/util/sector';
 import type { TypedEvent, TypedEventListener } from '../../util/observable';
 import { Observable } from '../../util/observable';
 import type { ChartAxis } from '../chartAxis';
@@ -1075,7 +1073,7 @@ export abstract class Series<
         throw new Error('AG Charts - Series.pickNodeMainAxisFirst() not implemented');
     }
 
-    public pickNodesInBBoxPredicate(): PickNodesInBBoxPredicate {
+    protected pickNodesInBBoxPredicate(): PickNodesInBBoxPredicate {
         // By default, pickNodesInBBox just used boxes for hit-testing because it's easier and faster. Series with more
         // complicated shapes (e.g. sectors or pie/donut, paths for maps) need to override this predicate to implement
         // their own hit-testing computation.
@@ -1084,19 +1082,8 @@ export abstract class Series<
         switch (containment) {
             case 'any':
                 return (selectionBox: BoxBounds, node: Node<unknown>): boolean => {
-                    if (node instanceof Sector) {
-                        const offset = Transformable.fromCanvasPoint(this.contentGroup, selectionBox.x, selectionBox.y);
-                        const seriesSelectionBox: BoxBounds = {
-                            x: offset.x,
-                            y: offset.y,
-                            width: selectionBox.width,
-                            height: selectionBox.height,
-                        };
-                        return boxCollidesSector(seriesSelectionBox, node);
-                    } else {
-                        const nodeBox = Transformable.toCanvas(this.contentGroup, node.getBBox());
-                        return boxCollides(selectionBox, nodeBox.x, nodeBox.y, nodeBox.width, nodeBox.height);
-                    }
+                    const nodeBox = Transformable.toCanvas(this.contentGroup, node.getBBox());
+                    return boxCollides(selectionBox, nodeBox.x, nodeBox.y, nodeBox.width, nodeBox.height);
                 };
             case 'all':
                 return (selectionBox: BoxBounds, node: Node<unknown>): boolean => {
